requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-cors-headers==1.1.0
django-countries==3.4.1
django-filter==0.11.0
django-oauth-toolkit==0.10.0
djangorestframework==3.3.2
djangorestframework-gis==0.10
drf-nested-routers==0.11.1
